Fix: Entity attributes aren't registered on servers, so they fail to spawn
